Practica8/requirements.txt
django==3.1.7
djangorestframework==3.12.2
mysqlclient==2.0.3
django-mysql==3.9
django-cors-headers==3.5.0
